+++ src/rust/src/lib.rs
@@ -3,6 +3,7 @@ use extendr_api::prelude::*;
 mod laser_frame;
 mod epidemic;
 mod distributions;
+mod migration;
 pub use laser_frame::LaserFrame;
 pub use distributions::Distribution;
 
@@ -11,4 +12,5 @@ extendr_module! {
     use laser_frame;
     use epidemic;
     use distributions;
+    use migration;
 }